refactor(server): extract the session state fold into session-fold
The fold (phase machine, active-turn buffer, pending requests, cursor)
was tangled with the registry that owns drain fibers. Split the pure
half out so a session's observable state stops being reachable only
through a live native stream.

Pure move plus a rename: Projection -> SessionState, fold ->
foldSessionEvent. No behaviour change.
